DSC-120, #3 - test in progress, update logic refactored (moved to a separate module), additional_families field updated to accept blanks
