+++ models.py
@@ -9,7 +9,6 @@ from webu.custom_models.registry import custom_model
 from cms.plugin_base import CMSPluginBase
 from cms.models.pluginmodel import CMSPlugin
 
-
 AUTH_USER_MODEL = settings.AUTH_USER_MODEL
 
 # Device Server statuses
@@ -148,12 +147,12 @@ class DscManagedModel(models.Model):
 
     # if object is changed it is marked as
     last_update_activity = models.ForeignKey('DeviceServerActivity',
-                                            related_name='updated_%(class)s',
-                                            default=None, null=True, blank=True)
+                                             related_name='updated_%(class)s',
+                                             default=None, null=True, blank=True)
 
     create_activity = models.ForeignKey('DeviceServerActivity',
-                                            related_name='created_%(class)s',
-                                            default=None, null=True, blank=True)
+                                        related_name='created_%(class)s',
+                                        default=None, null=True, blank=True)
 
     def is_valid(self):
         if self.invalidate_activity is None:
@@ -177,12 +176,12 @@ class DscManagedModel(models.Model):
 
     def last_update_method(self):
         create_object = None
-        if self.last_update_activity is not None and hasattr(self.last_update_activity,'create_object'):
+        if self.last_update_activity is not None and hasattr(self.last_update_activity, 'create_object'):
             create_object = self.last_update_activity.create_object
-        elif self.create_activity is not None and hasattr(self.create_activity,'create_object'):
+        elif self.create_activity is not None and hasattr(self.create_activity, 'create_object'):
             create_object = self.create_activity.create_object
 
-        if create_object is not None and create_object.all().count()>0:
+        if create_object is not None and create_object.all().count() > 0:
             if create_object.first().use_uploaded_xmi_file:
                 return 'file'
             elif create_object.first().use_url_xmi_file:
@@ -196,7 +195,7 @@ class DscManagedModel(models.Model):
         elif self.create_activity is not None and hasattr(self.create_activity, 'create_object'):
             create_object = self.create_activity.create_object
 
-        if create_object is not None and create_object.all().count()>0:
+        if create_object is not None and create_object.all().count() > 0:
             return create_object.first().script_operation
         return False
 
@@ -281,9 +280,7 @@ class DeviceServer(DscManagedModel):
             ('delete_own_deviceserver', 'deletion of device server for creator'),
         )
 
-
-    # TODO addverbose names
-
+    # TODO add verbose names
 
     #
     #
@@ -420,7 +417,7 @@ class DeviceServerDocumentation(DscManagedModel):
     device_server = models.ForeignKey(DeviceServer, related_name='documentation')
 
     def get_documentation_url(self):
-        if self.documentation_file is not None and hasattr(self.documentation_file,'url'):
+        if self.documentation_file is not None and hasattr(self.documentation_file, 'url'):
             return self.documentation_file.url
         else:
             return self.url
@@ -544,13 +541,13 @@ class DeviceClassInfo(DscManagedModel):
 
     product_reference = models.CharField(max_length=255, verbose_name="Product", default='')
 
-    additional_families = models.ManyToManyField(DeviceClassFamily, related_name='device_class_infos')
+    additional_families = models.ManyToManyField(DeviceClassFamily, related_name='device_class_infos', blank=True)
 
     def all_families(self):
-        families = [ self.class_family, ] + [dcf.family for dcf in self.additional_families]
+        families = [self.class_family, ] + [dcf.family for dcf in self.additional_families]
         return set(families)
 
-    def make_backup(self,activity, cl=None):
+    def make_backup(self, activity, cl=None):
         cli = DeviceClassInfo()
         if cl is None:
             cli.device_class = self.device_class
@@ -566,7 +563,7 @@ class DeviceClassInfo(DscManagedModel):
         cli.create_activity = self.create_activity
         cli.last_update_activity = self.last_update_activity
         cli.invalidate_activity = activity
-        cli.save() # need to have an id to upfate many-to-many field
+        cli.save()  # need to have an id to upfate many-to-many field
         cli.additional_families.add(*self.additional_families.all())
 
         return cli
@@ -612,13 +609,13 @@ class DeviceCommand(DscManagedModel):
 
     input_type = models.SlugField(choices=zip(DS_COMMAND_DATATYPES.values(),
                                               DS_COMMAND_DATATYPES.values()),
-                                  verbose_name='Argument Type',blank=True, null=True)
+                                  verbose_name='Argument Type', blank=True, null=True)
 
     input_description = models.TextField(verbose_name='Argin description', blank=True, null=True)
 
     output_type = models.SlugField(choices=zip(DS_COMMAND_DATATYPES.values(),
                                                DS_COMMAND_DATATYPES.values()),
-                                   verbose_name='Output Type',blank=True, null=True)
+                                   verbose_name='Output Type', blank=True, null=True)
 
     output_description = models.TextField(verbose_name='Argout description', blank=True, null=True)
 
@@ -676,6 +673,8 @@ class DeviceAttributeInfo(DscManagedModel):
         return '%s' % self.device_attribute
 
 ###############################################################################################
+
+from xmi_parser import TangoXmiParser
 
 
 class DeviceServerAddModel(models.Model):
@@ -736,7 +735,7 @@ class DeviceServerAddModel(models.Model):
 
     repository_contact = models.EmailField(verbose_name='Please write to', blank=True, null=True, default='')
 
-    repository_download_url= models.URLField(verbose_name='Dwonload URL', blank=True, default='')
+    repository_download_url = models.URLField(verbose_name='Dwonload URL', blank=True, default='')
 
     repository_tag = models.CharField(verbose_name='Tag', max_length=128, blank=True, null=True, default='')
 
@@ -823,15 +822,15 @@ class DeviceServerAddModel(models.Model):
                                            max_length=2000000
                                            )
 
-
     # xmi file
     use_uploaded_xmi_file = models.BooleanField(verbose_name='Upload .xmi file to populate database.',
                                                 blank=True,
                                                 default=True)
 
-    add_class = models.BooleanField(verbose_name='Add new class to the device server instead of updating existing ones?',
-                                                blank=True,
-                                                default=False)
+    add_class = models.BooleanField(
+        verbose_name='Add new class to the device server instead of updating existing ones?',
+        blank=True,
+        default=False)
 
     xmi_file = models.FileField(verbose_name='.XMI file',
                                 upload_to='dsc_xmi_files',
@@ -839,8 +838,8 @@ class DeviceServerAddModel(models.Model):
                                 null=True)
 
     use_url_xmi_file = models.BooleanField(verbose_name='Provide .xmi file URL populate database.',
-                                                blank=True,
-                                                default=False)
+                                           blank=True,
+                                           default=False)
 
     xmi_file_url = models.URLField(verbose_name='XMI source URL', blank=True, null=True, default='')
 
@@ -849,7 +848,7 @@ class DeviceServerAddModel(models.Model):
             :return  (xmi_string, is_ok, error_message)
         """
 
-        if hasattr(self,'valid_xmi_string'):
+        if hasattr(self, 'valid_xmi_string'):
             if self.valid_xmi_string is not None:
                 return self.valid_xmi_string, True, ''
         try:
@@ -902,7 +901,7 @@ class DeviceServerAddModel(models.Model):
 
     class_family = models.CharField(max_length=128, blank=True, null=True, default='')  # TODO: implement choices
 
-    additional_families = models.ManyToManyField(DeviceClassFamily)
+    additional_families = models.ManyToManyField(DeviceClassFamily, blank=True)
 
     platform = models.CharField(max_length=64,
                                 choices=zip(['Windows', 'Unix Like', 'All Platforms'],
@@ -1084,433 +1083,6 @@ def search_device_servers(search_text):
     qall = qn | qd | qm | qp | qf | qb | qk
 
     return qall.filter(invalidate_activity=None).distinct()
-
-
-from xmi_parser import TangoXmiParser
-
-
-def create_or_update(update_object, activity, device_server=None, device_class=None):
-    """this method creates or updates device server based on update/add object. It should be used inside atimic
-    transaction to keep database consistent.
-        :return (device_server, backup_device_server)
-    """
-
-    assert isinstance(update_object, DeviceServerAddModel)
-
-    backup_device_server = None
-    new_device_server = None
-
-    # backup device server if it is update
-    if device_server is not None:
-        backup_device_server = device_server.make_backup(activity=activity)
-        backup_device_server.save()
-
-    # basic information about device server
-    if update_object.use_uploaded_xmi_file or update_object.use_url_xmi_file:
-        # use xmi file
-        xmi_string, ok, message = update_object.xmi_string()
-        print '-------------'
-        print message
-        if ok:
-            parser = TangoXmiParser(xml_string=xmi_string)
-            # device server object
-            new_device_server = parser.get_device_server()
-
-    else:
-        new_device_server = DeviceServer(name='', description='')
-
-    if update_object.certified is not None:
-        new_device_server.certified = update_object.certified
-
-    if update_object.development_status is not None:
-        new_device_server.development_status = update_object.development_status
-
-    if update_object.ds_status is not None:
-        new_device_server.status = update_object.ds_status
-
-    if update_object.picture is not None:
-        new_device_server.picture = update_object.picture
-
-    if update_object.ds_info_copy:
-        new_device_server.description = ''
-        if not update_object.use_url_xmi_file and not update_object.use_uploaded_xmi_file \
-                and len(update_object.class_name) > 0:
-            new_device_server.name = update_object.class_name
-            if len(update_object.license_name) > 0:
-                lic = DeviceServerLicense.objects.get_or_create(name=update_object.license_name)[0]
-                if lic.pk is None:
-                    lic.save()
-                new_device_server.license = lic
-
-    # if data provided manually
-    if not update_object.ds_info_copy:
-        new_device_server.name = update_object.name
-        new_device_server.description = update_object.description
-
-    # make sure license object is saved
-    if new_device_server.license is not None and new_device_server.license.pk is None:
-        new_device_server.license.save()
-
-    # so, we have base information
-    if device_server is None:
-        device_server = new_device_server
-        device_server.create_activity = activity
-        device_server.created_by = activity.created_by
-        device_server.status = STATUS_NEW
-    else:
-        if not update_object.ds_info_copy or update_object.use_manual_info or update_object.use_url_xmi_file or \
-                update_object.use_uploaded_xmi_file or \
-                (update_object.ds_info_copy and new_device_server.name != device_server.name):
-            # the if above is to avoid update when not asked for
-            device_server.name = new_device_server.name
-            device_server.description = new_device_server.description
-            device_server.license = new_device_server.license
-        device_server.last_update_activity = activity
-        device_server.development_status = new_device_server.development_status
-        device_server.certified = new_device_server.certified
-        device_server.status = STATUS_UPDATED
-        if new_device_server.picture is not None:
-            device_server.picture = new_device_server.picture
-
-    # make sure device server has pk
-    device_server.save()
-
-    # # mark it in activity
-    # activity.device_server = device_server
-    # activity.device_server_backup = backup_device_server
-    # activity.save()
-
-    # check repository
-    new_repository = None
-    if update_object.available_in_repository or update_object.repository_contact is not None \
-            or update_object.repository_download_url is not None:
-        if not update_object.available_in_repository:
-            update_object.repository_type = None
-            update_object.repository_url = None
-            update_object.repository_path = None
-            update_object.repository_download_url = None
-            update_object.repository_tag = None
-        new_repository = DeviceServerRepository(repository_type=update_object.repository_type,
-                                                url=update_object.repository_url,
-                                                path_in_repository=update_object.repository_path,
-                                                contact_email=update_object.repository_contact,
-                                                download_url=update_object.repository_download_url,
-                                                tag=update_object.repository_tag
-                                                )
-        if backup_device_server is not None:
-            # if it is update operation and repository change old repo is redirected to backup device server
-            # and new_repository is marked as update
-            if hasattr(device_server, 'repository'):
-                assert isinstance(device_server.repository, DeviceServerRepository)
-                # check also if there is a real change
-                if new_repository.repository_type != device_server.repository.repository_type \
-                        or new_repository.url != device_server.repository.url \
-                        or new_repository.path_in_repository != device_server.repository.path_in_repository \
-                        or new_repository.download_url != device_server.repository.download_url \
-                        or new_repository.tag != device_server.repository.tag \
-                        or new_repository.contact_email != device_server.repository.contact_email:
-                    # for modified repository move it to backup
-                    old_repo = device_server.repository
-                    old_repo.device_server = backup_device_server
-                    old_repo.invalidate_activity = activity
-                    old_repo.save()
-                    new_repository.last_update_activity = activity
-                    new_repository.device_server = device_server
-                    new_repository.create_activity = old_repo.create_activity
-
-                else:
-                    # no real change
-                    new_repository = None
-        else:
-            # here, it means we are creating a new device server
-            new_repository.create_activity = activity
-            new_repository.device_server = device_server
-
-    # make sure the repository is saved
-    if new_repository is not None:
-        new_repository.last_update_activity = activity
-        new_repository.device_server = device_server
-
-        new_repository.save()
-
-    # check documentation
-    # readme is update directly
-    if update_object.upload_readme:
-        device_server.readme = update_object.readme_file
-
-    # old documentation will be
-    current_docs = device_server.documentation.filter(invalidate_activity=None)
-    doc = None
-
-    new_documentation = DeviceServerDocumentation(documentation_type=update_object.documentation1_type,
-                                                  url=update_object.documentation1_url,
-                                                  title=update_object.documentation1_title,
-                                                  documentation_file=update_object.documentation1_file,
-                                                  create_activity=activity,
-                                                  device_server=device_server)
-
-    if update_object.documentation1_pk is not None:
-        doc = current_docs.filter(pk=update_object.documentation1_pk).first()
-        if doc is not None and (not update_object.script_operation or doc.updated_by_script()):
-            if update_object.other_documentation1:
-
-                assert isinstance(doc, DeviceServerDocumentation)
-
-                if doc.documentation_type != update_object.documentation1_type \
-                        or update_object.documentation1_file is not None \
-                        or update_object.documentation1_url != doc.url \
-                        or update_object.documentation1_title != doc.title:
-
-                    new_documentation.save()
-
-                    if backup_device_server is not None:
-                        doc.device_server = backup_device_server
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-                    else:
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-
-            else:
-                if backup_device_server is not None:
-                    doc.device_server = backup_device_server
-                doc.make_invalid(activity=activity)
-                doc.save()
-
-    elif update_object.other_documentation1:
-        new_documentation.save()
-
-    # doc 2
-    doc = None
-
-    new_documentation = DeviceServerDocumentation(documentation_type=update_object.documentation2_type,
-                                                  url=update_object.documentation2_url,
-                                                  title=update_object.documentation2_title,
-                                                  documentation_file=update_object.documentation2_file,
-                                                  create_activity=activity,
-                                                  device_server=device_server)
-
-    if update_object.documentation2_pk is not None:
-        doc = current_docs.filter(pk=update_object.documentation2_pk).first()
-        if doc is not None and (not update_object.script_operation or doc.updated_by_script()):
-            if update_object.other_documentation2:
-
-                assert isinstance(doc, DeviceServerDocumentation)
-
-                if doc.documentation_type != update_object.documentation2_type \
-                        or update_object.documentation2_file is not None \
-                        or update_object.documentation2_url != doc.url \
-                        or update_object.documentation2_title != doc.title:
-
-                    new_documentation.save()
-
-                    if backup_device_server is not None:
-                        doc.device_server = backup_device_server
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-                    else:
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-
-            else:
-                if backup_device_server is not None:
-                    doc.device_server = backup_device_server
-                doc.make_invalid(activity=activity)
-                doc.save()
-
-    elif update_object.other_documentation2:
-        new_documentation.save()
-
-    # doc 3
-    doc = None
-
-    new_documentation = DeviceServerDocumentation(documentation_type=update_object.documentation3_type,
-                                                  url=update_object.documentation3_url,
-                                                  title=update_object.documentation3_title,
-                                                  documentation_file=update_object.documentation3_file,
-                                                  create_activity=activity,
-                                                  device_server=device_server)
-
-    if update_object.documentation3_pk is not None:
-        doc = current_docs.filter(pk=update_object.documentation3_pk).first()
-        if doc is not None and (not update_object.script_operation or doc.updated_by_script()):
-            if update_object.other_documentation1:
-
-                assert isinstance(doc, DeviceServerDocumentation)
-
-                if doc.documentation_type != update_object.documentation3_type \
-                        or update_object.documentation3_file is not None \
-                        or update_object.documentation3_url != doc.url \
-                        or update_object.documentation3_title != doc.title:
-
-                    new_documentation.save()
-
-                    if backup_device_server is not None:
-                        doc.device_server = backup_device_server
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-                    else:
-                        doc.make_invalid(activity=activity)
-                        doc.save()
-
-            else:
-                if backup_device_server is not None:
-                    doc.device_server = backup_device_server
-                doc.make_invalid(activity=activity)
-                doc.save()
-
-    elif update_object.other_documentation3:
-        new_documentation.save()
-
-
-            # get classes and interface
-    if (update_object.use_uploaded_xmi_file or update_object.use_url_xmi_file ) and parser is not None:
-        # for xmi file all old device classes and relation are moved to backup and new are created from scratch
-        if not update_object.add_class:
-            if device_class is None:
-                for cl in device_server.device_classes.all():
-                    cl.make_invalid(activity)
-                    if backup_device_server is not None:
-                        cl.device_server = backup_device_server
-                    cl.save()
-            else:
-                cl = device_server.device_classes.filter(pk=device_class).first()
-                cl.make_invalid(activity)
-                if backup_device_server is not None:
-                    cl.device_server = backup_device_server
-                cl.save()
-
-        # read classes from xmi file
-        cls = parser.get_device_classes()
-        for cl in cls:
-            # save class
-            assert (isinstance(cl, DeviceClass))
-            cl.device_server = device_server
-            cl.create_activity = activity
-            if backup_device_server is not None:
-                cl.last_update_activity = activity
-            if cl.license is not None and cl.license.pk is None:
-                cl.license.save()
-
-            cl.save()
-
-            # create and save class info object
-            class_info = parser.get_device_class_info(cl)
-            assert (isinstance(class_info, DeviceClassInfo))
-            class_info.device_class = cl
-            class_info.save() # need to do it to be able to populate m2m field
-            class_info.additional_families.clear()
-            class_info.additional_families.add(*update_object.additional_families.all())
-            class_info.create_activity = activity
-            if backup_device_server is not None:
-                class_info.last_update_activity = activity
-
-            # if manual info provided override defaults
-            if update_object.use_manual_info:
-                if len(update_object.contact_email) > 0:
-                    class_info.contact_email = update_object.contact_email
-
-            class_info.save()
-
-            # create and save attributes
-            attributes = parser.get_device_attributes(cl)
-            for attrib in attributes:
-                assert (isinstance(attrib, DeviceAttribute))
-                attrib.device_class = cl
-                attrib.create_activity = activity
-                attrib.save()
-
-            # create and save commands
-            commands = parser.get_device_commands(cl)
-            for command in commands:
-                assert (isinstance(command, DeviceCommand))
-                command.device_class = cl
-                command.create_activity = activity
-                command.save()
-
-            # create and save pipes
-            pipes = parser.get_device_pipes(cl)
-            for pipe in pipes:
-                assert (isinstance(pipe, DevicePipe))
-                pipe.device_class = cl
-                pipe.create_activity = activity
-                pipe.save()
-
-            # create and save properties
-            properties = parser.get_device_properties(cl)
-            for prop in properties:
-                assert (isinstance(prop, DeviceProperty))
-                prop.device_class = cl
-                prop.create_activity = activity
-                prop.save()
-
-    elif update_object.use_manual_info:
-        if backup_device_server is not None and not update_object.add_class:
-            if device_class is None:
-                clq = device_server.device_classes.all()
-            else:
-                clq = device_server.device_classes.filter(pk=device_class)
-
-            for cl in clq:
-                clo = cl.make_backup(activity)
-                clo.device_server = backup_device_server
-                clo.save()
-                if hasattr(cl,'info'):
-                    cloi = cl.info.make_backup(activity, clo)
-                    cloi.save()
-
-        # manual info provided
-        if device_class is None:
-            clq = device_server.device_classes.all()
-
-        else:
-            clq = device_server.device_classes.filter(pk=device_class)
-
-        if clq.count() > 0 and not update_object.add_class:
-            cl = clq.first()
-            cl.last_update_activity = activity
-            cl_info = cl.info
-            cl_info.last_update_activity = activity
-        else:
-            cl = DeviceClass()
-            cl.create_activity = activity
-            cl_info = DeviceClassInfo()
-            cl_info.create_activity = activity
-            if backup_device_server is not None:
-                cl.last_update_activity = activity
-                cl_info.last_update_activity = activity
-
-        cl.name = update_object.class_name
-        cl.description = update_object.class_description
-        cl.class_copyright = update_object.class_copyright
-        cl.language = update_object.language
-        cl.device_server = device_server
-
-        if len(update_object.license_name) > 0:
-            lic = DeviceServerLicense.objects.get_or_create(name=update_object.license_name)[0]
-            if lic.pk is None:
-                lic.save()
-            cl.license = lic
-
-        cl.save()
-
-        cl_info.device_class = cl
-        cl_info.contact_email = update_object.contact_email
-        cl_info.class_family = update_object.class_family
-        cl_info.platform = update_object.platform
-        cl_info.bus = update_object.bus
-        cl_info.manufacturer = update_object.manufacturer
-        cl_info.product_reference = update_object.product_reference
-        cl_info.key_words = update_object.key_words
-        cl_info.save() # make sure to that cl_inf has a valid id
-        cl_info.additional_families.clear()
-        cl_info.additional_families.add(*update_object.additional_families.all())
-
-        cl_info.save()
-
-    # return device_server
-    return device_server, backup_device_server
-
 
 ###################################################################################
